+++ proj/260210-stack-integrated/review/00-SUMMARY.md
@@ -2,7 +2,7 @@
 
 > **Status Update (2026-02-17, `pr-review-fix`):** The maintainability blocker from this snapshot is still partially remediated.
 > Import-cycle remediation and validation gates are green (`scripts/detect_cycles.py` reports no cycles, `mypy`/`make test-all`/remote docker target pass).
-> `tools/packages.py` was split and reduced below 1000 LOC (758), and the production TODO in that module was removed.
+> `tools/packages.py` was split and reduced below 1000 LOC (758), `tools/responses.py` was reduced below 1000 LOC (987), and the production TODO in `tools/packages.py` was removed.
 > Large-module structural debt still remains in other files (>1000 LOC).
 
 **Date:** 2026-02-17
