+++ proj/260210-stack-integrated/review/11-module-cleanup.md
@@ -1,8 +1,8 @@
 # 11 - Module Cleanup Tasklist
 
 > **Status Note (2026-02-17):** This workstream is actively in progress on `pr-review-fix`.
-> Recent completion: package CRUD/browse/diff/create/update/delete logic moved out of `tools/packages.py` into `tools/package_crud.py`, reducing `tools/packages.py` to 758 lines and keeping validation green (`make test-all`, `make test-remote-docker`, `mypy`, cycle checks).
-> Remaining debt: several modules are still >1000 lines (`ops/quilt_ops.py`, `services/governance_service.py`, `backends/platform_admin_ops.py`, `backends/platform_backend.py`, `services/workflow_service.py`, `tools/responses.py`).
+> Recent completion: package CRUD/browse/diff/create/update/delete logic moved out of `tools/packages.py` into `tools/package_crud.py` (now 758 LOC), and response base/resource models were extracted from `tools/responses.py` (now 987 LOC).
+> Remaining debt: several modules are still >1000 lines (`ops/quilt_ops.py`, `services/governance_service.py`, `backends/platform_admin_ops.py`, `backends/platform_backend.py`, `services/workflow_service.py`).
 
 **Date:** 2026-02-16
 **Reviewer:** Codex
